--- louis/agent/louis.py
@@ -1,113 +1,0 @@
-import os
-import openai
-
-def nonewlines(s: str) -> str:
-    return s.replace('\n', ' ').replace('\r', ' ')
-
-AZURE_OPENAI_SERVICE = os.environ.get("AZURE_OPENAI_SERVICE") or "myopenai"
-AZURE_OPENAI_GPT_DEPLOYMENT = os.environ.get("AZURE_OPENAI_GPT_DEPLOYMENT") or "davinci"
-AZURE_OPENAI_CHATGPT_DEPLOYMENT = os.environ.get("AZURE_OPENAI_CHATGPT_DEPLOYMENT") or "chat"
-
-class Louis:
-    prompt_prefix = """<|im_start|>system
-You are Louis, a french and english bilingual virtual assistant for the Canadian Food Inspection Agency.
-The Canadian Food Inspection Agency (CFIA) is a regulatory agency that is responsible for safeguarding Canada's food supply
-and protecting the health of Canadians.
-The CFIA's mandate includes a wide range of activities related to food safety, animal health, and plant health,
-including inspection, testing, and enforcement.
-Answer in the langage of the question asked and in both french and english when specifically requested.
-Answer ONLY with the facts listed in the list of sources below. If there isn't enough information below, say you don't know.
-Do not generate answers that don't use the sources below.
-If asking a clarifying question to the user would help, ask the question.
-For tabular information return it as an html table. Do not return markdown format.
-Each source has a name followed by colon and the actual information, always include the source name for each fact you use in the response.
-Use square brakets to reference the source, e.g. [info1.txt].
-Don't combine sources, list each source separately, e.g. [info1.txt][info2.pdf].
-{follow_up_questions_prompt}
-{injected_prompt}
-Sources:
-{sources}
-<|im_end|>
-{chat_history}
-"""
-
-    follow_up_questions_prompt_content = """Generate three very brief follow-up questions that the user would likely ask next about laws and regulations related to the agency.
-    Use double angle brackets to reference the questions, e.g. <<What regulations do I need to be aware of?>>.
-    Try not to repeat questions that have already been asked.
-    Only generate questions and do not generate any text before or after the questions, such as 'Next Questions'"""
-
-    query_prompt_template = """Below is a history of the conversation so far, and a new question asked by the user that needs to be answered by searching in a knowledge base.
-    Generate a search query based on the conversation and the new question.
-    Do not include cited source filenames and document names e.g info.txt or doc.pdf in the search query terms.
-    Do not include any text inside [] or <<>> in the search query terms.
-    If the question is not in English, translate the question to English before generating the search query.
-
-Chat History:
-{chat_history}
-
-Question:
-{question}
-
-Search query:
-"""
-
-    def __init__(self, search_client, chatgpt_deployment=AZURE_OPENAI_CHATGPT_DEPLOYMENT, gpt_deployment=AZURE_OPENAI_GPT_DEPLOYMENT):
-        self.search_client = search_client
-        self.chatgpt_deployment = chatgpt_deployment
-        self.gpt_deployment = gpt_deployment
-
-    def run(self, history: list[dict], overrides: dict) -> any:
-        use_semantic_captions = True if overrides.get("semantic_captions") else False
-        top = overrides.get("top") or 3
-        exclude_category = overrides.get("exclude_category") or None
-        filter = "category ne '{}'".format(exclude_category.replace("'", "''")) if exclude_category else None
-
-        # STEP 1: Generate an optimized keyword search query based on the chat history and the last question
-        prompt = self.query_prompt_template.format(chat_history=self.get_chat_history_as_text(history, include_last_turn=False), question=history[-1]["user"])
-        completion = openai.Completion.create(
-            engine=self.gpt_deployment,
-            prompt=prompt,
-            temperature=0.0,
-            max_tokens=32,
-            n=1,
-            stop=["\n"])
-        q = completion.choices[0].text
-
-        results = self.search_client(q)
-        print(results)
-        content = nonewlines("\n".join([r['content'] for r in results]))
-
-        follow_up_questions_prompt = self.follow_up_questions_prompt_content if overrides.get("suggest_followup_questions") else ""
-
-        # Allow client to replace the entire prompt, or to inject into the exiting prompt using >>>
-        prompt_override = overrides.get("prompt_template")
-        if prompt_override is None:
-            prompt = self.prompt_prefix.format(injected_prompt="", sources=content, chat_history=self.get_chat_history_as_text(history), follow_up_questions_prompt=follow_up_questions_prompt)
-        elif prompt_override.startswith(">>>"):
-            prompt = self.prompt_prefix.format(injected_prompt=prompt_override[3:] + "\n", sources=content, chat_history=self.get_chat_history_as_text(history), follow_up_questions_prompt=follow_up_questions_prompt)
-        else:
-            prompt = prompt_override.format(sources=content, chat_history=self.get_chat_history_as_text(history), follow_up_questions_prompt=follow_up_questions_prompt)
-
-        # STEP 3: Generate a contextual and content specific answer using the search results and chat history
-        completion = openai.Completion.create(
-            engine=self.chatgpt_deployment,
-            prompt=prompt,
-            temperature=overrides.get("temperature") or 0.7,
-            max_tokens=1024,
-            n=1,
-            stop=["<|im_end|>", "<|im_start|>"])
-
-        retvalue = {
-            "data_points": content,
-            "answer": completion.choices[0].text,
-            "thoughts": f"Searched for:<br>{q}<br><br>Prompt:<br>" + prompt.replace('\n', '<br>')
-        }
-        return retvalue
-
-    def get_chat_history_as_text(self, history, include_last_turn=True, approx_max_tokens=1000) -> str:
-        history_text = ""
-        for h in reversed(history if include_last_turn else history[:-1]):
-            history_text = """<|im_start|>user""" +"\n" + h["user"] + "\n" + """<|im_end|>""" + "\n" + """<|im_start|>assistant""" + "\n" + (h.get("bot") + """<|im_end|>""" if h.get("bot") else "") + "\n" + history_text
-            if len(history_text) > approx_max_tokens*4:
-                break
-        return history_text
