+++ louis/blueprints/chat.py
@@ -2,12 +2,12 @@ import logging
 
 from flask import Blueprint, request, jsonify
 
-from louis.agent.chain import ChainAgent
-from louis.agent.louis import Louis
-from louis.actions import smartsearch
+from louis.agents.chain import ChainAgent
+from louis.agents.louis import Louis
+from louis.tools.smartsearch import SmartSearch
 
 chat = Blueprint('chat', __name__)
-louis = Louis(smartsearch)
+louis = Louis(SmartSearch)
 
 @chat.route("/", methods=["POST"])
 def chat_user():
