--- louis/blueprints/static.py
@@ -1,3 +1,0 @@
-from flask import Blueprint
-static = Blueprint('static', __name__)
-
