+++ louis/prompts/clarification.txt
@@ -1,0 +1,1 @@
+If asking a clarifying question to the user would help, ask the question.
