+++ louis/prompts/follow_up_questions_prompt.txt
@@ -1,0 +1,6 @@
+Generate three very brief follow-up questions that the user would likely ask
+next about laws and regulations related to the agency. Use double angle brackets
+to reference the questions, e.g. <<What regulations do I need to be aware of?>>.
+Try not to repeat questions that have already been asked. Only generate
+questions and do not generate any text before or after the questions, such as
+'Next Questions'
