+++ louis/prompts/format.txt
@@ -1,0 +1,6 @@
+
+For tabular information return it as an html table. Do not return markdown
+format. Each source has a name followed by colon and the actual information,
+always include the source name for each fact you use in the response. Use square
+brakets to reference the source, e.g. [info1.txt]. Don't combine sources, list
+each source separately, e.g. [info1.txt][info2.pdf].
