+++ louis/prompts/query_prompt_template.txt
@@ -1,0 +1,14 @@
+Below is a history of the conversation so far, and a new question asked by the
+user that needs to be answered by searching in a knowledge base. Generate a
+search query based on the conversation and the new question. Do not include
+cited source filenames and document names e.g info.txt or doc.pdf in the search
+query terms. Do not include any text inside [] or <<>> in the search query
+terms.
+
+Chat History:
+{chat_history}
+
+Question:
+{question}
+
+Search query:
